test(graph): close failed trace review gaps
